+++ utils/util.py
@@ -15,3 +15,6 @@ def setup_config(title):
         layout="wide",
         initial_sidebar_state="expanded"
     )
+
+def get_brand_colors_list():
+    return ['#d4bc64', '#84ccb4', '#5c74b4', '#6c6c84', '#504f8f', '#83c3d4', '#646c84', '#646c7c', '#588898', '#586c9c']