+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- [Code Cleanup](https://trello.com/c/JfXHt5tB/7-limpieza-y-ajustes-de-codigo)
+
 ### Added
-- Demo menu e inicio de sesión.
-- Reportes y 360 del jugador
+- [Login](https://trello.com/c/TOqsAhot/1-login)
+    - Login and menu
+- [Player Reports](https://trello.com/c/LdtXga0f/4-dise%C3%B1ar-layout-visual-para-ficha-de-jugador)
+    - Player reports and 360 dashboard
+- [Player Evaluation](https://trello.com/c/4nMRTyJg/9-a%C3%B1adir-formulario-de-registro-de-jugadores)